Migrate connection related keywords to ConnectionKeywords
