+++ Mainframe3270/keywords/__init__.py
@@ -1,0 +1,1 @@
+from Mainframe3270.keywords.connection import ConnectionKeywords  # noqa: F401
